Refactor byte_string
