+++ byte_string/src/lib.rs
@@ -1,11 +1,10 @@
 use std::borrow::Cow;
-use std::fmt::{self, Display};
 
 mod from_bytes;
 use from_bytes::from_bytes;
 pub use from_bytes::{Number, ParseIntError};
 
-#[derive(Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub struct ByteStr<'inner> {
     bytes: &'inner [u8],
 }
@@ -52,27 +51,6 @@ impl<'outer> ByteStr<'outer> {
     }
 }
 
-impl<'outer, T> From<&'outer T> for ByteStr<'outer>
-where
-    T: AsRef<[u8]> + ?Sized
-{
-    fn from(other: &'outer T) -> Self {
-        ByteStr { bytes: other.as_ref() }
-    }
-}
-
-impl AsRef<[u8]> for ByteStr<'_> {
-    fn as_ref(&self) -> &[u8] {
-        self.bytes
-    }
-}
-
-impl Display for ByteStr<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        write!(f, "{}", self.to_str_lossy())
-    }
-}
-
 #[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
 pub struct ByteString {
     bytes: Vec<u8>,
@@ -104,235 +82,97 @@ impl ByteString {
     }
 }
 
-impl From<Vec<u8>> for ByteString
-{
-    fn from(other: Vec<u8>) -> Self {
-        ByteString { bytes: other }
-    }
-}
-
-impl From<String> for ByteString {
-    fn from(other: String) -> Self {
-        ByteString { bytes: other.into_bytes() }
-    }
-}
-
-impl<T> From<&T> for ByteString
-where
-    T: AsRef<[u8]> + ?Sized
-{
-    fn from(other: &T) -> Self {
-        ByteString { bytes: other.as_ref().to_vec() }
-    }
-}
-
-impl<'a> std::ops::Deref for ByteStr<'a> {
-    type Target = &'a [u8];
-
-    fn deref(&self) -> &Self::Target {
-        &self.bytes
-    }
-}
-
-impl std::ops::Deref for ByteString {
-    type Target = Vec<u8>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.bytes
-    }
-}
-
-impl std::ops::DerefMut for ByteString {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.bytes
-    }
-}
-
-impl AsRef<[u8]> for ByteString {
-    fn as_ref(&self) -> &[u8] {
-        self.bytes.as_ref()
-    }
-}
-
-impl Display for ByteString {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        self.as_byte_str().fmt(f)
-    }
-}
-
-#[cfg(test)]
-mod tests {
+mod impl_from {
     use super::*;
 
-    #[test]
-    fn test_byte_str_size() {
-        assert_eq!(std::mem::size_of::<&[u8]>(), 16);
-        assert_eq!(std::mem::size_of::<ByteStr>(), 16);
-    }
-    #[test]
-    fn test_byte_str_new() {
-        let a = ByteStr::new(b"hello");
-        let b = ByteStr::new("hello");
-        assert_eq!(a, b);
+    impl<'outer, T> From<&'outer T> for ByteStr<'outer>
+    where
+        T: AsRef<[u8]> + ?Sized
+    {
+        fn from(other: &'outer T) -> Self {
+            ByteStr { bytes: other.as_ref() }
+        }
     }
 
-    #[test]
-    fn test_byte_str_from() {
-        let a = ByteStr::from(b"hello");
-        let b = ByteStr::from("hello");
-        assert_eq!(a, b);
+    impl From<Vec<u8>> for ByteString
+    {
+        fn from(other: Vec<u8>) -> Self {
+            ByteString { bytes: other }
+        }
     }
 
-    #[test]
-    fn test_byte_str_to_str_lossy() {
-        let a = ByteStr::from("hello");
-        assert_eq!(a.to_str_lossy(), "hello");
+    impl From<String> for ByteString {
+        fn from(other: String) -> Self {
+            ByteString { bytes: other.into_bytes() }
+        }
     }
 
-    #[test]
-    fn test_byte_str_display() {
-        use std::fmt::Write;
+    impl<T> From<&T> for ByteString
+    where
+        T: AsRef<[u8]> + ?Sized
+    {
+        fn from(other: &T) -> Self {
+            ByteString { bytes: other.as_ref().to_vec() }
+        }
+    }
+}
 
-        let a = ByteStr::from("hello");
-        let mut buf = String::new();
+mod impl_as_ref {
+    use super::*;
 
-        write!(buf, "{}", a).unwrap();
-
-        assert_eq!(buf, "hello")
+    impl AsRef<[u8]> for ByteStr<'_> {
+        fn as_ref(&self) -> &[u8] {
+            self.bytes
+        }
     }
 
-    #[test]
-    fn test_byte_str_eq_ignore_ascii_case() {
-        let a = ByteStr::from("hEllo");
-        let b = ByteStr::from("helLo");
+    impl AsRef<[u8]> for ByteString {
+        fn as_ref(&self) -> &[u8] {
+            self.bytes.as_ref()
+        }
+    }
+}
 
-        assert!(a.eq_ignore_ascii_case(&b));
-        assert!(a.eq_ignore_ascii_case(b"HeLlO"));
-        assert!(a.eq_ignore_ascii_case("HeLlO"));
+mod impl_display {
+    use super::*;
+    use std::fmt::{self, Display};
+
+    impl Display for ByteStr<'_> {
+        fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+            write!(f, "{}", self.to_str_lossy())
+        }
     }
 
-    #[test]
-    fn test_byte_string() {
-        let a = ByteString::new();
-        assert_eq!(a.len(), 0);
+    impl Display for ByteString {
+        fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+            self.as_byte_str().fmt(f)
+        }
+    }
+}
+
+mod impl_deref {
+    use super::*;
+    use std::ops::{Deref, DerefMut};
+
+    impl<'a> Deref for ByteStr<'a> {
+        type Target = &'a [u8];
+
+        fn deref(&self) -> &Self::Target {
+            &self.bytes
+        }
     }
 
-    #[test]
-    fn test_byte_string_from() {
-        // Vecs and Strings are moved
-        let _a = ByteString::from(b"hello".to_vec());
-        let _b = ByteString::from("hello".to_string());
+    impl Deref for ByteString {
+        type Target = Vec<u8>;
 
-        // Slice/reference types are cloned
-        let _c = ByteString::from(b"hello");
-        let _d = ByteString::from("hello");
+        fn deref(&self) -> &Self::Target {
+            &self.bytes
+        }
     }
 
-    #[test]
-    fn test_byte_string_display() {
-        use std::fmt::Write;
-
-        let a = ByteString::from("hello");
-        let mut buf = String::new();
-
-        write!(buf, "{}", a).unwrap();
-
-        assert_eq!(buf, "hello")
-    }
-
-    #[test]
-    fn test_byte_str_deref() {
-        let a = ByteStr::from(b"hello");
-        assert_eq!(a.len(), 5);
-    }
-
-    #[test]
-    fn test_byte_string_deref() {
-        let a = ByteString::from(b"hello");
-        assert_eq!(a.len(), 5);
-    }
-
-    #[test]
-    fn test_byte_string_mut_deref() {
-        let mut a = ByteString::from(b"hello");
-        a.push(b'a');
-        assert_eq!(a.len(), 6);
-    }
-
-    #[test]
-    fn test_byte_string_as_byte_str() {
-        let a = ByteString::from(b"hello");
-        let _b: ByteStr = a.as_byte_str();
-    }
-
-    #[test]
-    fn test_byte_string_into_vec() {
-        fn assert_vec(_: Vec<u8>) {};
-
-        let a = ByteString::from(b"hello");
-        assert_vec(a.into_vec());
-    }
-
-    #[test]
-    fn test_byte_string_as_ref_u8() {
-        fn assert_as_ref(_: impl AsRef<[u8]>) {};
-
-        let a = ByteString::from(b"hello");
-        assert_as_ref(&a);
-        assert_eq!(a.as_ref(), b"hello");
-    }
-
-    #[test]
-    fn test_byte_string_as_hashmap_key() {
-        use std::collections::HashMap;
-        let mut h = HashMap::new();
-        h.insert(ByteString::from("a"), 1);
-    }
-
-    #[test]
-    fn test_byte_string_clone() {
-        let a: ByteString = "a".into();
-        let b = a.clone();
-        assert_eq!(a, b);
-    }
-
-    #[test]
-    fn test_byte_str_parse() {
-        assert_eq!(ByteStr::from(b"1").parse::<i64>(), Ok(1i64));
-        assert_eq!(ByteStr::from(b"x").parse::<i64>(), Err(ParseIntError));
-    }
-
-    #[test]
-    fn test_byte_string_parse() {
-        assert_eq!(ByteString::from(b"1").parse::<i64>(), Ok(1i64));
-        assert_eq!(ByteString::from(b"x").parse::<i64>(), Err(ParseIntError));
-    }
-
-    #[test]
-    fn test_byte_str_to_lowercase() {
-        let a: ByteStr = "abcABC123\x01".into();
-        let b = a.to_lowercase();
-        assert_eq!(b, "abcabc123\x01".into())
-    }
-
-    #[test]
-    fn test_byte_string_to_lowercase() {
-        let a: ByteString = "abcABC123\x01".into();
-        let b = a.to_lowercase();
-        assert_eq!(b, "abcabc123\x01".into())
-    }
-
-    #[test]
-    fn test_byte_str_to_uppercase() {
-        let a: ByteStr = "abcABC123\x01".into();
-        let b = a.to_uppercase();
-        assert_eq!(b, "ABCABC123\x01".into())
-    }
-
-    #[test]
-    fn test_byte_string_to_uppercase() {
-        let a: ByteString = "abcABC123\x01".into();
-        let b = a.to_uppercase();
-        assert_eq!(b, "ABCABC123\x01".into())
+    impl DerefMut for ByteString {
+        fn deref_mut(&mut self) -> &mut Self::Target {
+            &mut self.bytes
+        }
     }
 }
